物物网管理平台

+++ mywuwu-visual/mywuwu-shop/src/main/java/com/mywuwu/pigx/shop/entity/Cart.java
@@ -79,12 +79,12 @@ private static final long serialVersionUID = 1L;
      * 市场价
      */
     @ApiModelProperty(value="市场价")
-    private BigDecimal marketPrice;
+    private Double marketPrice;
     /**
      * 零售价格
      */
     @ApiModelProperty(value="零售价格")
-    private BigDecimal retailPrice;
+    private Double retailPrice;
     /**
      * 数量
      */
